Fix Home and Landing pages
